\name{(INTRO.DISTRSIM)}
\alias{(INTRO.DISTRSIM)}
\title{DISTRSIM}
\center{Simulation Classes for Package distr}
\synopsis{}

\section{Usage}{
   library("distrSim") 
}


\description{
\pkg{``distrSim''} is a package for R from version 2.2.0 onwards that is distributed
under GPL license 2.0. It requires package \pkg{setRNG} by Paul Gilbert, \email{pgilbert@bank-banque-canada.ca},
and \pkg{distr} by the authors of this package, \email{peter.ruckdeschel@uni-bayreuth.de},
to be installed from 

\url{http://cran.r-project.org/mirrors.html}

The aim of this package is to provide a conceptual treatment of simulations by means of S4 classes.
The package is based on our package \pkg{distr}, hence uses distribution classes 
as introduced there to describe the distributions under which simulations are performed.

A somewhat longer and more detailed manual can be obtained under

\url{http://www.uni-bayreuth.de/departments/math/org/mathe7/DISTR/}

}
\keyword{}
