\name{Data.c-methods}
\docType{methods}
\alias{Data.c-methods}
\alias{Data.c}
\alias{Data.c,Contsimulation-method}
\title{ Methods for Function Data.c in Package `distrSim' }
\description{Data.c-methods}

\section{Methods}{\describe{
\item{Data.c}{\code{signature(object = "Dataclass")}: returns the
  contaminated data }
}}
\keyword{}
