\name{Data.id-methods}
\docType{methods}
\alias{Data.id-methods}
\alias{Data.id}
\alias{Data.id,Contsimulation-method}
\title{ Methods for Function Data.id in Package `distrSim'}
\description{Data.id-methods}

\section{Methods}{\describe{

\item{Data.id}{\code{signature(object = "Contsimulation")}: returns the ideal data }
}}
\keyword{}
