\name{distribution.c-methods}
\docType{methods}
\alias{distribution.c-methods}
\alias{distribution.c<--methods}
\alias{distribution.c}
\alias{distribution.c<-}
\alias{distribution.c,Contsimulation-method}
\alias{distribution.c<-,Contsimulation-method}
\title{ Methods for Function distribution.c in Package `distrSim' }
\description{distribution-methods}

\section{Methods}{\describe{

\item{distribution.c}{\code{signature(object = "Contsimulation")}: returns the distribution of the contaminated data}
\item{distribution.c<-}{\code{signature(object = "Contsimulation")}: changes the distribution of the contaminated data}
}}
\concept{simulation}
\concept{S4 simulation class}
\concept{accessor function}
\concept{replacement function}
\concept{contaminating distribution}
\concept{contamination}
\keyword{methods}
\keyword{distribution}
