\name{distribution.id-methods}
\docType{methods}
\alias{distribution.id-methods}
\alias{distribution.id<--methods}
\alias{distribution.id}
\alias{distribution.id<-}
\alias{distribution.id,Contsimulation-method}
\alias{distribution.id<-,Contsimulation-method}
\title{ Methods for Function distribution.id in Package `distrSim' }
\description{distribution-methods}
\section{Methods}{\describe{
\item{distribution.id}{\code{signature(object = "Contsimulation")}: returns the distribution of the ideal data }
\item{distribution.id<-}{\code{signature(object = "Contsimulation")}: changes the distribution of the ideal data}
}}
\keyword{}
