\name{filename-methods}
\docType{methods}
\alias{filename-methods}
\alias{filename<--methods}
\alias{filename}
\alias{filename<-}
\alias{filename,Dataclass-method}
\alias{filename<-,Dataclass-method}

\title{ Methods for Function filename in Package `distrSim' }
\description{filename-methods}
\section{Methods}{\describe{

\item{filename}{\code{signature(object = "Dataclass")}: returns the filename }

\item{filename<-}{\code{signature(object = "Dataclass")}: changes the filename }
    
}}
\keyword{file}
\keyword{manip}
\keyword{methods}
\concept{S4 data class}
\concept{accessor function}
\concept{replacement function}
