\name{ind-methods}
\docType{methods}
\alias{ind-methods}
\alias{ind}
\alias{ind,Contsimulation-method}
\title{ Methods for Function ind in Package `distrSim'}
\description{ind-methods}

\section{Methods}{\describe{

\item{ind}{\code{signature(object = "Contsimulation")}: returns an indicator showing which data is contaminated }
}}
\keyword{manip}
\concept{accessor function}
\keyword{methods}
\concept{S4 data class}
