\name{obsDim-methods}
\docType{methods}
\alias{obsDim-methods}
\alias{obsDim}
\alias{obsDim,Dataclass-method}
\title{Methods for Function obsDim in Package `distrSim'}
\description{obsDim-methods}

\section{Methods}{\describe{

\item{obsDim}{\code{signature(object = "Dataclass")}: returns the number of obsDim }
}}
\keyword{}
