\name{plot-methods}
\docType{methods}
\alias{plot-methods}
\alias{plot,Dataclass-method}
\alias{plot,Simulation-method}
\alias{plot,Contsimulation-method}
\title{ Methods for Function plot in Package `distrSim' }
\description{plot-methods}
\section{Methods}{\describe{
\item{plot}{\code{signature(object = "Dataclass")}: produces a plot of the data matrix }
\item{plot}{\code{signature(object = "Simulation")}: produces a plot of the data matrix }
\item{plot}{\code{signature(object = "Contsimulation")}: produces a plot of the real data matrix }
}}
\keyword{}
