\name{print-methods}
\docType{methods}
\alias{print-methods}
\alias{print,Dataclass-method}
\alias{print,Simulation-method}
\alias{print,Contsimulation-method}
\title{ Methods for Function print in Package `distrSim' }
\description{print-methods}

\section{Methods}{\describe{


\item{print}{\code{signature(x = "Dataclass")}: returns filename, number of runs and the size of the sample }

\item{print}{\code{signature(x = "Simulation")}: returns filename, seed, number of runs, the size of the sample and the distribution }

\item{print}{\code{signature(x = "Contsimulation")}: returns filename, seed, number of runs, the size of the sample, the rate and the distributions }

}}
\keyword{}
