\name{runs-methods}
\docType{methods}
\alias{runs-methods}
\alias{runs<--methods}
\alias{runs}
\alias{runs<-}
\alias{runs,Dataclass-method}
\alias{runs<-,Simulation-method}
\alias{runs<-,Contsimulation-method}
\title{ Methods for Function runs in Package `distrSim'}
\description{runs-methods}

\section{Methods}{\describe{

\item{runs}{\code{signature(object = "Dataclass")}: returns the number of runs }
\item{runs<-}{\code{signature(object = "Simulation")}: changes the number of runs }
\item{runs<-}{\code{signature(object = "Contsimulation")}: changes the number of runs }
}}
\keyword{}
