\name{samplesize-methods}
\docType{methods}
\alias{samplesize-methods}
\alias{samplesize<--methods}
\alias{samplesize}
\alias{samplesize<-}
\alias{samplesize,Dataclass-method}
\alias{samplesize<-,Simulation-method}
\alias{samplesize<-,Contsimulation-method}
\title{ Methods for Function samplesize in Package `distrSim'}
\description{samplesize-methods}

\section{Methods}{\describe{

\item{samplesize}{\code{signature(object = "Dataclass")}: returns the size of the sample }
\item{samplesize<-}{\code{signature(object = "Simulation")}: changes the size of the sample }
\item{samplesize<-}{\code{signature(object = "Contsimulation")}: changes the size of the sample }
}}
\keyword{methods}
\concept{simulation}
\concept{S4 simulation class}
\concept{pseudo random numbers}
\concept{RNG}
\concept{sample}
\concept{sample size}
\concept{accessor function}
\concept{replacement function}
