\name{simulate-methods}
\docType{methods}
\alias{simulate-methods}
\alias{simulate}
\alias{simulate,Simulation-method}
\alias{simulate,Contsimulation-method}
\title{ Methods for Function simulate in Package `distrSim' }
\description{simulate-methods}

\section{Methods}{\describe{

\item{simulate}{\code{signature(object = "Simulation")}: generates the random numbers for the simulation }

\item{simulate}{\code{signature(object = "Contsimulation")}: generates the random numbers for the simulation }

}}
\keyword{methods}
\concept{simulation}
\concept{S4 simulation class}
\concept{pseudo random numbers}
\concept{RNG}
\keyword{distribution}