\name{vectororNULL-class}
\docType{class}
\alias{vectororNULL-class}

\title{Class "vectororNULL" }
\description{ auxiliary class; may contain either a vector or NULL, cf. J. Chambers, "green book".}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "vectororNULL" in the signature.
}
\author{Thomas Stabla \email{Thomas.Stabla@uni-bayreuth.de},\cr Florian Camphausen \email{Florian.Camphausen@uni-bayreuth.de},\cr
 Peter Ruckdeschel \email{Peter.Ruckdeschel@uni-bayreuth.de},\cr Matthias Kohl \email{Matthias.Kohl@stamats.de}}
\note{ Dataclass-class can save data either of type "NULL" (means no data) or "vector"}
\code{\link{Dataclass-class}}
\keyword{}
