\name{obsDim-methods}
\docType{methods}
\alias{obsDim-methods}
\alias{obsDim}
\alias{obsDim,Dataclass-method}
\alias{obsDim,SeqDataFrames-method}
\title{Methods for Function obsDim in Package `distrSim'}
\description{obsDim-methods}

\section{Methods}{\describe{

\item{obsDim}{\code{signature(object = "Dataclass")} or \code{signature(object = "SeqDataFrames")}: returns the number of obsDim }
}}
\keyword{manip}
\concept{accessor function}
\keyword{methods}
\concept{S4 data class}
