\name{rate-methods}
\docType{methods}
\alias{rate-methods}
\alias{rate<--methods}
\alias{rate}
\alias{rate<-}
\alias{rate,Contsimulation-method}
\alias{rate<-,Contsimulation-method}
\title{ Methods for Function rate in Package `distrSim' }
\description{rate-methods}

\section{Methods}{\describe{

\item{rate}{\code{signature(object = "Contsimulation")}: returns the contamination rate }
\item{rate<-}{\code{signature(object = "Contsimulation")}: modifies the contamination rate }
}}
\keyword{methods}
\concept{simulation}
\concept{S4 simulation class}
\concept{contamination}
\concept{contamination rate}
\concept{accessor function}
\concept{replacement function}
